saving things

+++ .backup_file.sh
@@ -3,7 +3,12 @@
 folders='documents literature'
 folders='documents music pictures literature'
 dir=/d
-dest=/media/oney/Vault
+
+if [ -d /media/oney/stuff/ ]; then
+    dest=/media/oney/stuff
+elif [ -d /media/oney/OneyCryptex/ ]; then
+    dest=/media/oney/Vault
+fi
 
 cd ~
 
@@ -14,23 +19,27 @@ if [ 2 -eq `date +%w` ]; then
     rsync -vurtl --delete --size-only /stuff/vms/ $dest/vms/
 fi
 
-for f in ${folders}
-do
-    # cp -r -u ${dir}/${f} ${dest}
-    # rsync -vurtCl ${dir}/${f}/ ${dest}/${f}
-    rsync -vurtl --delete ${dir}/${f}/ ${dest}/${f}/
-done
+if [ -n $dest ]; then
+    for f in ${folders}; do
+        rsync -vurtl --delete ${dir}/${f}/ ${dest}/${f}/
+    done
+fi
 
 # for the cryptex
 if [ -d /media/oney/OneyCryptex/ ]; then
     echo "Everything looks good...";
     rsync -vurtCl --delete /d/documents/ /media/oney/OneyCryptex/documents/
     rsync -vurtl --delete ~ /media/oney/OneyCryptex/
-    rsync -vurtRCl --delete .icedove/ /media/oney/OneyCryptex/
-else echo "mount the partition first..."; exit;
+    rsync -vurtRCl --delete .thunderbird/ /media/oney/OneyCryptex/
+else echo "mount the partition first...";
 fi
 
-# Try to keep everything
-# f=Videos
-# rsync -vurtl ${dir}/${f}/ ${dest}/${f}/
-# rsync -vurtl ${dir}/${f}/Movies/watchedMovies/ ${dest}/old/watchedMovies/
+if [ -d /media/oney/stuff/ ]; then
+    echo "Everything looks good...";
+    rsync -vurtl --delete ~ $dest/
+    rsync -vurtRCl --delete .thunderbird/ $dest/
+    # for f in ${folders}; do
+    #     rsync -vurtl --delete ${dir}/${f}/ ${dest}/${f}/
+    # done
+else echo "mount the partition first..."; 
+fi
